Move async strategies to annotation support
