+++ instrumentation/guava-10.0/library/build.gradle.kts
@@ -4,4 +4,5 @@ plugins {
 
 dependencies {
   library("com.google.guava:guava:10.0")
+  implementation(project(":instrumentation-api-annotation-support"))
 }
