+++ instrumentation/reactor-3.1/library/build.gradle.kts
@@ -4,6 +4,7 @@ plugins {
 
 dependencies {
   library("io.projectreactor:reactor-core:3.1.0.RELEASE")
+  implementation(project(":instrumentation-api-annotation-support"))
   testLibrary("io.projectreactor:reactor-test:3.1.0.RELEASE")
 
   testImplementation(project(":instrumentation:reactor-3.1:testing"))
