+++ instrumentation/rxjava/rxjava-3.0/library/build.gradle.kts
@@ -4,6 +4,7 @@ plugins {
 
 dependencies {
   library("io.reactivex.rxjava3:rxjava:3.0.12")
+  implementation(project(":instrumentation-api-annotation-support"))
 
   testImplementation(project(":instrumentation:rxjava:rxjava-3.0:testing"))
 }
