+++ javaagent-tooling/build.gradle.kts
@@ -10,6 +10,7 @@ dependencies {
   implementation(project(":javaagent-extension-api"))
   implementation(project(":javaagent-instrumentation-api"))
   implementation(project(":instrumentation-api"))
+  implementation(project(":instrumentation-api-annotation-support"))
   implementation(project(":muzzle"))
 
   implementation("io.opentelemetry:opentelemetry-api")
